odb: use code generator for new iterator and add      add reverse logic to it
